deployment-script/deployment db/requirements.txt
pymongo==4.2.0
kafka-python==2.0.2